created lib header which contains the functions
